refactor: code review fixes — dedup proxy, fetch-or-404, async error handling, Promise.all, Decimal type, ENV guard

+++ product-service/src/schemas.py
@@ -1,31 +1,29 @@
 from datetime import datetime
+from decimal import Decimal
 from pydantic import BaseModel, ConfigDict
 
 
-# What the client sends when creating a product
 class ProductCreate(BaseModel):
     name: str
     description: str | None = None
-    price: float
+    price: Decimal
     stock: int = 0
 
 
-# What the client sends when updating a product (all fields optional)
 class ProductUpdate(BaseModel):
     name: str | None = None
     description: str | None = None
-    price: float | None = None
+    price: Decimal | None = None
     stock: int | None = None
 
 
-# What we send back in responses — includes DB-generated fields
 class ProductResponse(BaseModel):
-    model_config = ConfigDict(from_attributes=True)  # lets Pydantic read SQLAlchemy models
+    model_config = ConfigDict(from_attributes=True)
 
     id: int
     name: str
     description: str | None
-    price: float
+    price: Decimal  # matches Numeric(10,2) in DB — avoids float precision loss
     stock: int
     created_at: datetime
     updated_at: datetime
